requirements.txt
matplotlib
pandas
numpy
seaborn
joblib
tqdm
umap-learn
scikit-learn
torch
datasets
